+++ src/output/index.js
@@ -22,6 +22,9 @@ module.exports = {
 
     console.log(chalk.bold('commands:'))
     console.log(`new project: ${chalk.green('websn projectName')}`)
+    console.log(
+      `new project with defaults: ${chalk.green('websn projectName -y')}`
+    )
     console.log(`help: ${chalk.green('websn --help')}`)
     console.log(`version: ${chalk.green('websn --version')}`)
 
@@ -47,11 +50,11 @@ module.exports = {
   },
   printVersion () {
     this.printNewLine()
-    console.log(`${chalk.bold('version:')} ${chalk.blue('1.2.1')}`)
+    console.log(`${chalk.bold('version:')} ${chalk.blue('1.3.0')}`)
   },
   printError (message) {
     this.printNewLine()
-    console.log(`${chalk.bold.red('error:')} ${chalk.red(message)}`)
+    console.error(`${chalk.bold.red('error:')} ${chalk.red(message)}`)
   },
   printSuccess (message) {
     this.printNewLine()