+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.3.0] - 2018-05-24
+
+### Added
+
++ Added flag for using default options when generating a new project
++ Added project generation tests
+
+### Changed
+
++ Optimized check for disallowed project names
++ Use tables instead of lists for displaying options in readme
++ Use yargs-parser instead of minimist since minimist is no longer maintained
++ Updated dependencies
+
 ## [1.2.1] - 2018-05-22
 
 ### Fixed
@@ -40,7 +54,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 + Initial release
 
-[Unreleased]: https://github.com/mserajnik/websn/compare/1.2.1...develop
+[Unreleased]: https://github.com/mserajnik/websn/compare/1.3.0...develop
+[1.3.0]: https://github.com/mserajnik/websn/compare/1.2.1...1.3.0
 [1.2.1]: https://github.com/mserajnik/websn/compare/1.2.0...1.2.1
 [1.2.0]: https://github.com/mserajnik/websn/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/mserajnik/websn/compare/1.0.0...1.1.0